live_trading_system/requirements.txt
oandapyV20
streamlit
pandas
python-dotenv
pytest
torch
scikit-learn
